# docs/api-doc.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "STR — Registration Number Prefill API",
    "version": "1.0.0",
    "description": "Vanjski handoff endpoint za zahtjev za registracijskim brojem (RB). Vanjski portali (TuStart i sl.) proslijeđuju autenticirane podatke najmoprimca, backend ih pohranjuje i vraća 302 redirect na frontend STR-a s neprovidnim `prefill` UUID-em u query stringu."
  },
  "servers": [
    { "url": "https://str.eturizam.gov.hr", "description": "Production (TBD — točan URL nije potvrđen; trenutni eTurizam: https://eturizam.gov.hr/)" },
    { "url": "https://str.testeduturizam.gov.hr", "description": "Test (CDU)" },
    { "url": "https://s-str-02.infodom.hr:8080", "description": "Dev (interni Infodom)" },
    { "url": "https://strbackend-production.up.railway.app", "description": "Mock (Railway)" }
  ],
  "tags": [
    { "name": "Registration Number Prefill", "description": "Handoff prefill payloada za zahtjev za RB." }
  ],
  "paths": {
    "/api/registration-number-prefill": {
      "get": {
        "tags": ["Registration Number Prefill"],
        "summary": "Handoff iz vanjskog portala",
        "description": "Sprema prefill payload i vraća 302 Found redirect na frontend formu STR-a. `Location` header sadrži `?prefill={uuid}` koji STR frontend interno razmjenjuje za pohranjene podatke.",
        "operationId": "handoff",
        "parameters": [
          {
            "name": "oib",
            "in": "query",
            "required": true,
            "description": "OIB najmoprimca, 11 znamenki.",
            "schema": { "type": "string", "minLength": 11, "maxLength": 11, "pattern": "^\\d{11}$", "example": "12345678901" }
          },
          {
            "name": "firstName",
            "in": "query",
            "required": true,
            "description": "Ime najmoprimca.",
            "schema": { "type": "string", "maxLength": 128, "example": "Ana" }
          },
          {
            "name": "lastName",
            "in": "query",
            "required": true,
            "description": "Prezime najmoprimca.",
            "schema": { "type": "string", "maxLength": 128, "example": "Anić" }
          },
          {
            "name": "addressCode",
            "in": "query",
            "required": false,
            "description": "Šifra kućnog broja iz vanjskog adresnog registra",
            "schema": { "type": "string", "maxLength": 64, "example": "42" }
          },
          {
            "name": "maxBedCount",
            "in": "query",
            "required": false,
            "description": "Broj kreveta u objektu.",
            "schema": { "type": "integer", "format": "int32", "minimum": 0, "example": 3 }
          },
          {
            "name": "maxGuestCount",
            "in": "query",
            "required": false,
            "description": "Maksimalni broj gostiju u objektu.",
            "schema": { "type": "integer", "format": "int32", "minimum": 0, "example": 6 }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect na frontend; payload je spremljen, UUID u `Location` headeru.",
            "headers": {
              "Location": {
                "description": "Frontend URL s `prefill` query parametrom.",
                "schema": { "type": "string", "example": "https://str-eturizam.gov.hr/registration-number?prefill=4e9c2c6a-2d3e-4a14-8a8f-1c5b5a8e5b71" }
              }
            }
          },
          "400": {
            "description": "Nedostaje obavezan parametar (`oib`, `firstName`, `lastName`) ili je format neispravan.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetail" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProblemDetail": {
        "type": "object",
        "description": "RFC 7807 problem details payload.",
        "properties": {
          "type": { "type": "string", "format": "uri", "example": "about:blank" },
          "title": { "type": "string", "example": "Bad Request" },
          "status": { "type": "integer", "format": "int32", "example": 400 },
          "detail": { "type": "string", "example": "Required request parameter 'oib' is not present" },
          "instance": { "type": "string", "format": "uri", "example": "/api/registration-number-prefill" }
        }
      }
    }
  }
}